+++ {{ cookiecutter.app_name }}/tests/test_cli.py
@@ -35,5 +35,6 @@ def test_version_runner(runner: CliRunner) -> None:
     result = runner.invoke(cli, ["--version"])
     assert result.exit_code == 0
     assert (
-        result.output == f"cli, version {metadata.version('{{ cookiecutter.app_name }}')}\n"
+        result.output
+        == f"cli, version {metadata.version('{{ cookiecutter.app_name }}')}\n"
     )